# pkg/query/flamegraph_arrow: Unsafe cast strings to bytes for equality check