It seems there was a regression in the implementation recently and running the benchmarks again has shown the `string` to `[]byte` casting to be the problem. This fixes the issue.

The new unsafe.StringData was added in Go 1.20.

flame graph table vs arrow on the main branch:
```
name      old time/op    new time/op    delta
Query-24    30.7ms ±10%    33.1ms ± 6%   +7.97%  (p=0.002 n=10+10)

name      old alloc/op   new alloc/op   delta
Query-24    16.4MB ± 1%    20.8MB ± 0%  +26.79%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Query-24      254k ± 0%      213k ± 0%  -16.01%  (p=0.000 n=9+10)
```
flame graph table vs arrow from this PR:
```
name      old time/op    new time/op    delta
Query-24    30.7ms ±10%    28.5ms ± 8%   -7.17%  (p=0.011 n=10+10)

name      old alloc/op   new alloc/op   delta
Query-24    16.4MB ± 1%    14.7MB ± 1%  -10.42%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Query-24      254k ± 0%      127k ± 1%  -49.90%  (p=0.000 n=9+10)
```

flame graph arrow `main` vs this PR.
```
name      old time/op    new time/op    delta
Query-24    33.1ms ± 6%    28.5ms ± 8%  -14.02%  (p=0.000 n=10+10)

name      old alloc/op   new alloc/op   delta
Query-24    20.8MB ± 0%    14.7MB ± 1%  -29.35%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Query-24      213k ± 0%      127k ± 1%  -40.35%  (p=0.000 n=10+10)
```